+++ packages/epub_view/test/interpreter_test.dart
@@ -10,9 +10,9 @@ import 'package:epub_view/src/epub_cfi/interpreter.dart';
 import 'package:html/dom.dart';
 
 Future<Uint8List> _loadTestBook() async {
-  final url = Directory.current.path.replaceFirst(
-      RegExp(r'/epub_view.*'), '/epub_view/test/assets/book.epub');
-  final file = File(url);
+  // final url = Directory.current.path.replaceFirst(
+  //     RegExp(r'/epub_view.*'), '/epub_view/test/assets/book.epub');
+  final file = File('test/assets/book.epub');
   return file.readAsBytes();
 }
 
@@ -32,7 +32,7 @@ void main() {
       final document =
           EpubCfiReader().chapterDocument(_book.Chapters![0].SubChapters![2]);
       result = EpubCfiInterpreter().searchLocalPathForHref(
-        document.documentElement,
+        document!.documentElement,
         _cfiFragment.path!.localPath!,
       );
     } catch (e) {
@@ -49,7 +49,7 @@ void main() {
     final document =
         EpubCfiReader().chapterDocument(_book.Chapters![0].SubChapters![1]);
     final result = EpubCfiInterpreter().searchLocalPathForHref(
-      document.documentElement,
+      document!.documentElement,
       _cfiFragment.path!.localPath!,
     )!;