+++ packages/explorer/lib/explorer.dart
@@ -1,7 +1,7 @@
 library explorer;
 
 export 'src/data/models/entry.dart';
-export 'src/data/navigator.dart';
+export 'src/data/provider.dart';
 export 'src/explorer.dart';
 export 'src/i18n/localization_delegate.dart';
 export 'src/ui/widgets/action.dart';