+++ packages/explorer/lib/src/i18n/messages/messages_en.dart
@@ -8,6 +8,8 @@
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
 // ignore_for_file:unused_import, file_names
+// ignore_for_file: type_annotate_public_apis
+// ignore_for_file: always_declare_return_types
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -19,21 +21,22 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
-  final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "actionCopyHere" : MessageLookupByLibrary.simpleMessage("Copy here"),
-    "actionMenuCopy" : MessageLookupByLibrary.simpleMessage("Copy"),
-    "actionMenuCut" : MessageLookupByLibrary.simpleMessage("Cut"),
-    "actionMenuDelete" : MessageLookupByLibrary.simpleMessage("Delete"),
-    "actionMenuOpen" : MessageLookupByLibrary.simpleMessage("Open"),
-    "actionMoveHere" : MessageLookupByLibrary.simpleMessage("Move here"),
-    "cancel" : MessageLookupByLibrary.simpleMessage("Cancel"),
-    "create" : MessageLookupByLibrary.simpleMessage("Create"),
-    "empty" : MessageLookupByLibrary.simpleMessage("Empty"),
-    "fileName" : MessageLookupByLibrary.simpleMessage("File name"),
-    "folderName" : MessageLookupByLibrary.simpleMessage("Folder name"),
-    "newFile" : MessageLookupByLibrary.simpleMessage("New file"),
-    "newFolder" : MessageLookupByLibrary.simpleMessage("New folder"),
-    "uploadFiles" : MessageLookupByLibrary.simpleMessage("Upload files")
-  };
+  final Map<String, dynamic> messages =
+      _notInlinedMessages(_notInlinedMessages);
+  static _notInlinedMessages(_) => <String, Function>{
+        "actionCopyHere": MessageLookupByLibrary.simpleMessage("Copy here"),
+        "actionMenuCopy": MessageLookupByLibrary.simpleMessage("Copy"),
+        "actionMenuCut": MessageLookupByLibrary.simpleMessage("Cut"),
+        "actionMenuDelete": MessageLookupByLibrary.simpleMessage("Delete"),
+        "actionMenuOpen": MessageLookupByLibrary.simpleMessage("Open"),
+        "actionMoveHere": MessageLookupByLibrary.simpleMessage("Move here"),
+        "cancel": MessageLookupByLibrary.simpleMessage("Cancel"),
+        "create": MessageLookupByLibrary.simpleMessage("Create"),
+        "empty": MessageLookupByLibrary.simpleMessage("Empty"),
+        "fileName": MessageLookupByLibrary.simpleMessage("File name"),
+        "folderName": MessageLookupByLibrary.simpleMessage("Folder name"),
+        "newFile": MessageLookupByLibrary.simpleMessage("New file"),
+        "newFolder": MessageLookupByLibrary.simpleMessage("New folder"),
+        "uploadFiles": MessageLookupByLibrary.simpleMessage("Upload files")
+      };
 }