+++ packages/explorer/lib/src/i18n/messages/messages_ru.dart
@@ -8,6 +8,8 @@
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
 // ignore_for_file:unused_import, file_names
+// ignore_for_file: type_annotate_public_apis
+// ignore_for_file: always_declare_return_types
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -19,21 +21,23 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'ru';
 
-  final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "actionCopyHere" : MessageLookupByLibrary.simpleMessage("Копировать сюда"),
-    "actionMenuCopy" : MessageLookupByLibrary.simpleMessage("Копировать"),
-    "actionMenuCut" : MessageLookupByLibrary.simpleMessage("Вырезать"),
-    "actionMenuDelete" : MessageLookupByLibrary.simpleMessage("Удалить"),
-    "actionMenuOpen" : MessageLookupByLibrary.simpleMessage("Открыть"),
-    "actionMoveHere" : MessageLookupByLibrary.simpleMessage("Вставить сюда"),
-    "cancel" : MessageLookupByLibrary.simpleMessage("Отменить"),
-    "create" : MessageLookupByLibrary.simpleMessage("Создать"),
-    "empty" : MessageLookupByLibrary.simpleMessage("Пусто"),
-    "fileName" : MessageLookupByLibrary.simpleMessage("Имя файла"),
-    "folderName" : MessageLookupByLibrary.simpleMessage("Имя папки"),
-    "newFile" : MessageLookupByLibrary.simpleMessage("Новый файл"),
-    "newFolder" : MessageLookupByLibrary.simpleMessage("Новая папка"),
-    "uploadFiles" : MessageLookupByLibrary.simpleMessage("Выгрузить файлы")
-  };
+  final Map<String, dynamic> messages =
+      _notInlinedMessages(_notInlinedMessages);
+  static _notInlinedMessages(_) => <String, Function>{
+        "actionCopyHere":
+            MessageLookupByLibrary.simpleMessage("Копировать сюда"),
+        "actionMenuCopy": MessageLookupByLibrary.simpleMessage("Копировать"),
+        "actionMenuCut": MessageLookupByLibrary.simpleMessage("Вырезать"),
+        "actionMenuDelete": MessageLookupByLibrary.simpleMessage("Удалить"),
+        "actionMenuOpen": MessageLookupByLibrary.simpleMessage("Открыть"),
+        "actionMoveHere": MessageLookupByLibrary.simpleMessage("Вставить сюда"),
+        "cancel": MessageLookupByLibrary.simpleMessage("Отменить"),
+        "create": MessageLookupByLibrary.simpleMessage("Создать"),
+        "empty": MessageLookupByLibrary.simpleMessage("Пусто"),
+        "fileName": MessageLookupByLibrary.simpleMessage("Имя файла"),
+        "folderName": MessageLookupByLibrary.simpleMessage("Имя папки"),
+        "newFile": MessageLookupByLibrary.simpleMessage("Новый файл"),
+        "newFolder": MessageLookupByLibrary.simpleMessage("Новая папка"),
+        "uploadFiles": MessageLookupByLibrary.simpleMessage("Выгрузить файлы")
+      };
 }