+++ packages/explorer/lib/src/ui/provider.dart
@@ -3,23 +3,23 @@ import 'package:flutter/widgets.dart';
 
 class ControllerProvider extends InheritedWidget {
   const ControllerProvider({
-    @required this.explorerController,
-    @required this.scrollController,
-    Widget child,
-    Key key,
+    required this.explorerController,
+    required this.scrollController,
+    required Widget child,
+    Key? key,
   }) : super(
           key: key,
           child: child,
         );
 
   final ExplorerController explorerController;
-  final ScrollController scrollController;
+  final ScrollController? scrollController;
 
   @override
   bool updateShouldNotify(ControllerProvider old) =>
       explorerController != old.explorerController ||
       scrollController != old.scrollController;
 
-  static ControllerProvider of(BuildContext context) =>
+  static ControllerProvider? of(BuildContext context) =>
       context.findAncestorWidgetOfExactType<ControllerProvider>();
 }