+++ packages/explorer/CHANGELOG.md
@@ -1,15 +1,30 @@
-## [1.1.2]
+## 2.0.0
+### Changed
+* Null-safety migration
+
+## 1.3.0
+### Changed
+* `uploadFiles` and `filePressed` callbacks moved from `ExplorerController` to `Explorer` widget 
+* Renamed `NavigatorExplorer -> ExplorerProvider` and `IoNavigatorExplorer -> IoExplorerProvider`
+### Added
+* More docs and some screenshots
+
+## 1.2.0
+
+* Show context menu on mouse right button clicked
+
+## 1.1.2
 
 * Fixed bread crumbs in windows
 
-## [1.1.1]
+## 1.1.1
 
 * Append i18n
 
-## [1.1.0]
+## 1.1.0
 
 * Added i18n
 
-## [1.0.0] 
+## 1.0.0 
 
 * Initial release