+++ packages/native_pdf_view/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.0.1
+
+* Fixed [issue#170](https://github.com/rbcprolabs/packages.flutter/issues/170), 4.0.0 not works
+
 ## 4.0.0
 
 * Null safety migration & Flutter v2 capability
@@ -5,6 +9,8 @@
 * Package `[extended_image]` replaced by `[photo_view]`
 * Break changes at `pageBuilder` callback (`PDFViewPageBuilder`)
 * Added `<BoxDecoration> backgroundDecoration` property at `PdfView`
+* Added optional `initialPage` property in `PdfController.loadDocument()`
+* Temporarily disabled `pageSnapping` option
 
 ## 3.9.1